backend/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.1
pymongo==4.6.2
flask-cors==4.0.0
redis==5.0.1
Flask-Limiter==3.5.0
celery==5.3.6
PyJWT==2.8.0
bcrypt==4.1.2
google-auth==2.28.1
